Use hash instead of raw passphrase in the marker.
