+++ sys/kern/init_main.c
@@ -711,93 +711,7 @@ SYSCTL_INT(_kern, OID_AUTO, init_shutdown_timeout,
 	CTLFLAG_RW, &init_shutdown_timeout, 0, "Shutdown timeout of init(8). "
 	"Unused within kernel, but used to control init(8)");
 
-
-#define MAX_DIGEST_BUFFER 1024
-#include <sys/stat.h>
-#include <sys/fcntl.h>
-#include <sys/uio.h>
-#include <sys/iov.h>
-
-
-static int mypanic(const char *msg) {
-	const int panic_reboot_wait_time = 10;
-	int loop;
-
-	printf("%s\n", msg);
-	for (loop = panic_reboot_wait_time * 10; loop > 0; --loop) {
-		DELAY(1000 * 100); /* 1/10th second */
-	}
-	panic("%s\n", msg);
-}
-
-
-static void tpm2_check_passphrase_marker() {
-	struct thread *td = curthread;
-
-	int error;
-	struct stat sb;
-	int fd;
-	struct iovec aiov;
-	struct uio auio;
-	char buf[MAX_DIGEST_BUFFER];
-
-	char *was_retrieved = kern_getenv("kern.geom.eli.passphrase.from_tpm2.was_retrieved");
-	char *passphrase = kern_getenv("kern.geom.eli.passphrase.from_tpm2.passphrase");
-
-	if (was_retrieved == NULL || was_retrieved[0] != '1') {
-		printf("Passphrase from TPM was not used - OK.\n");
-		return;
-	}
-
-	if (passphrase == NULL) {
-		panic("Passphrase was retrieved from the TPM but was not passed to us.\n");
-	}
-
-	error = kern_statat(td, 0, AT_FDCWD, "/.passphrase_marker", UIO_SYSSPACE, &sb, NULL);
-	if (error) {
-		mypanic("kern_statat() on passphrase marker failed");
-	}
-
-	if (sb.st_mode & 0077) {
-		mypanic("Passphrase marker has wrong permissions set");
-	}
-
-	if (sb.st_size >= MAX_DIGEST_BUFFER) {
-		mypanic("Passphrase marker too long");
-	}
-
-	error = kern_openat(td, AT_FDCWD, "/.passphrase_marker", UIO_SYSSPACE, O_RDONLY, 0);
-	if (error) {
-		mypanic("Cannot open the passphrase marker");
-	}
-	fd = td->td_retval[0];
-	printf("fd: %d\n", fd);
-
-	aiov.iov_base = &buf[0];
-	aiov.iov_len = sb.st_size;
-	auio.uio_iov = &aiov;
-	auio.uio_iovcnt = 1;
-	auio.uio_resid = sb.st_size;
-	auio.uio_segflg = UIO_SYSSPACE;
-	error = kern_readv(td, fd, &auio);
-	if (error) {
-		mypanic("Failed to read the passphrase marker");
-	}
-	buf[sb.st_size] = '\0';
-
-	if (strncmp(buf, passphrase, MAX_DIGEST_BUFFER) != 0) {
-		mypanic("Passphrase marker does not match");
-	}
-
-	printf("Passphrase marker found and matching - we are done.\n");
-	kern_unsetenv("kern.geom.eli.passphrase.from_tpm2.passphrase");
-
-	error = kern_close(td, fd);
-	if (error) {
-		printf("Failed to close passphrase marker - that's weird.\n");
-	}
-}
-
+void tpm2_check_passphrase_marker(void);
 
 /*
  * Start the initial user process; try exec'ing each pathname in init_path.
